slack-bot/requirements.txt
kubernetes==29.0.0
slack-machine==0.36.0